Allow the mspdb DLL to be found via vswhere.exe
The incredibly useful `vswhere.exe` tool can be used to abstract away
all the nitty-gritty details of figuring out where Visual Studio is
installed.

This patch adds support for making use of that tool.

Since spawning a process is pretty expensive compared to looking at the
registry, we only fall back on this tool (if it exists) if the other
methods to find the DLL have failed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>