Verify in the GitHub workflow that cv2pdb can handle GCC's output
Git for Windows uses cv2pdb to convert the debugging information
produced by GCC to `.pdb` format, for use with Visual C's tools.

A recent GCC upgrade broke this because the debugging information was
produced in DWARF5 format, and the Git for Windows maintainer had to
scramble to implement (minimal) support for that format.

Let's verify in the GitHub workflow that cv2pdb can handle GCC's output
format, and let's use the very same GCC version that Git for Windows
uses to verify that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>