Add support for anon types in FQN generation
This fixes a bug in the fully qualified name generation we encounter a
segment that has no name, such as when a proc is really a C++ lambda
expression, which internally generates an anonymous class to house the
function. In this case, we generate a unique name based on the DIE's
offset to serve as the segment's name.