Fix bug in DWARF v4 location decoding
This fixes a long standing bug in the DWARF v4 location decoding logic.
Previously, the logic would not recognize "base address selection"
entries at all and assume an entry was either an end-of-list or a
location list entry thus resulting in incorrect decoding in release
builds, and an assert in debug builds.

See DWARF v4 spec section 2.6.2 for details.