Support namespaces/nesting for structs
This commit extends the FQN generation for procs to
structs/classes/unions using a similar pattern of looking up the parent
chain during construction of the name.

Like procs, we need to we need to perform a similar "deduplication" and
only emit a single entry for each struct. As with procs, DWARF has
multiple "flavors" of a single struct as multiple DIE records. All but
one are marked as a "declaration" and lack member information.  We want
the one with member information in the PDB as that will allow the
debugger to inspect the contents of objects of that type.

What makes structs tricker than procs is some formal parameters, local
(stack) variables, etc. reference the "declaration" flavors of the DIE,
which lack the member information. This makes this complicated for two
reasons:

1) We don't want to emit the declaration versions at all due to the
   consolidation reasons explained earlier.
2) If we don't emit the declaration versions, we need to "patch up" the
   type references of those params/locals to point to the definition
   entry we emit.

But (2) in itself is tricky, because, alas, DWARF has no direct pointer
from a "declaration" entry of a struct to is corresponding definition,
like it does with proc (procs have a "DW_AT_specification" attribute
which points from decl -> def. See 2.13.2 for details).

Alas, what I resorted to is a (hopefully precise) heuristic. I do the
matching manually by:

1) Introducing an index that lets me look up a DWARF entry by its
   (local, unqualified) name.
2) Using that index, look at all the matches (ignoring the decls, of
   course) and walking up their parent chains and comparing the tags and
   names of their ancestors.
3) If all of those things check out, we can be reasonably confident that
   this candidate entry is the logically equivalent "definition" entry
   for the "declaration" struct entry we're trying to resolve the typeID
   for.

So far it's worked well on the binaries I've tested it on. With this, we
can now dump structures (including nested and namespaced) ones in
WinDbg.